Debug (float64)

+++ internal/model/allmodels.go
@@ -13,26 +13,26 @@ type Order struct {
 	ID         int       `json:"-"`
 	Number     int       `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    int       `json:"accrual,omitempty"`
+	Accrual    float64   `json:"accrual,omitempty"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
 
 type Balance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
 type Withdrawals struct {
 	ID             int       `json:"-"`
 	NewOrderNumber string    `json:"order"`
-	Sum            int       `json:"sum"`
+	Sum            float64   `json:"sum"`
 	ProcessedAt    time.Time `json:"processed_at"`
 }
 
 type AccrualSystem struct {
-	OrderNumber string `json:"order"`
-	Status      string `json:"status"`
-	Accrual     int    `json:"accrual"`
-	RetryAfter  int    `json:"-"`
-	Message     string `json:"-"`
+	OrderNumber string  `json:"order"`
+	Status      string  `json:"status"`
+	Accrual     float64 `json:"accrual"`
+	RetryAfter  int     `json:"-"`
+	Message     string  `json:"-"`
 }
